Minimal support for unions in all supported bingings.
This adds the minimal support for unions in the bindings generation so that
generation itself does not fail. In particular, the generated code is not
correct.

This allows to start working on the union generation on each languages as
long as the mojom using unions are only used in languages that support
these.

In particular, this CL adds the tests for union in C++ in the build.

R=azani@chromium.org, johnmccutchan@google.com, john@johnmccutchan.com, rogulenko@google.com

Review URL: https://codereview.chromium.org/1128003006

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: f6b56e31e8dcb05240a1eb03dcb9134df9d64688