Add the ability to specify deps on zip android action.
Forward the zip action as a public dependency from the
mojo_android_application. This allows the dependencies to be set up to specify
inputs to this action, and to use the outputs in other actions.

Hide the zip action behind visibility.

R=sky@chromium.org

Review URL: https://codereview.chromium.org/1162673003

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 55ddc6aed64ba059bd36f598686bd946d56c3334