Add a roll_versions.py script and roll versions once.
This isn't perfect, since it makes non-minimal modifications
to the yaml and xml files.  In the yaml case it sorts the keys
and in the xml case it doesn't do the fancy attribute wrapping,
but this seems good enough to use.

R=abarth@chromium.org, johnmccutchen@google.com

Review URL: https://codereview.chromium.org/1179633003.

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 1b7bcee09d2ee4fa784741499cdcc95ca4a14ff9