go/bindings: handle messages with 0 request ID correctly
Previously Go code assumed that the request id in the message header
couldn't be 0. But it turned out that dart send messages with 0
request id. So this CL updates the Go code to be able to talk to dart.

R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/1189113006.

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 28310f303557cc82fb6969c73152212f014d6c4f