Dart: Allows the mojom package's generate script to download .mojoms.
This change adds an option to generate.dart that causes it to look for
files named '.mojoms' in a package. That file contains https:// URLs
for .mojom files. The generate.dart script downloads the files and
invokes Mojo's bindings generation on them.

This change also fixes a but in Dart bindings generation, which
caused a failure when generating only bindings for Dart.

BUG=
R=johnmccutchan@google.com

Review URL: https://codereview.chromium.org/1210253006 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: f92948b9155cb3429e69b12fdecbe479aa79bc7a