Make Message responsible for allocating and owning message data
This removes the AdoptData method. MessageBuilder then has to create a Message, and so it gains one as a member. FixedBuffer is changed to no longer allocate memory as MessageBuilder instead passes a pointer to the Message owned data buffer to FixedBuffer. FixedBufferForTesting is added to support unit tests that were using the old FixedBuffer directly.

This change paves the way for improving the way we allocate message data. For example, Message could have a small buffer as a member variable, which would have the effect of using the stack to allocate message data. Or, if the Mojo system API grew a function to allocate a message buffer we could use that as well.

R=yzshen@chromium.org

Review URL: https://codereview.chromium.org/1248963004 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 8eba9846e02a395b463f400b9ef8d08a9734e064