Basics for python script that can generate bindings.
The goal is to create a python script which can take a Mojom describing
a set of mojom files and generate the associated bindings.

This CL is the set of BUILD files and python path hacks necessary to
import all the necessary pieces.

BUG=
R=rudominer@chromium.org

Review URL: https://codereview.chromium.org/1347813007 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 218b7ea77ee7b958f46e90148d9e589704c671a7