This is the latest in a series of attempts to find the right way to expose Mojo service descriptions. This replaces the previous attempt at https://codereview.chromium.org/1374483002/edit.
Here we are bailing on the idea of using control messages. Instead the client first gets a hold of the ServiceDescriber corresponding to a ServiceProvider by invoking ServiceProvider.ConnectToService("_ServiceDescriber").

Then the client gets a ServiceDescription for a named service by invoking ServiceDescriber.DescribeService(interface_name).

The comments in this file introduce the new terminology of the *top-level interface* for a service and the *complete type set* of a service and make it clear that the scope of a ServiceDescription is the complete type set.

R=alexfandrianto@google.com, azani@chromium.org, alexfandrianto, azani
BUG=#448

Review URL: https://codereview.chromium.org/1381023005 .

Cr-Mirrored-From: https://github.com/domokit/mojo
Cr-Mirrored-Commit: 38bf3749048d447ae0905f3134cbda68f11f1ed2